# [IMP] sale_order_product_assortment: performance in so