This commit improves performance when computing the products allowed in a Sales Order, especially when creating an order and the partner is empty