@CarlosRoca13 2024-12-09
Hi @aleuffre, first of all thanks for the changes :smile: can you show a comparison of the times between the current version and after adding these changes?

ping @carlosdauden