@aleuffre 2024-12-09
I tried using `timeit` but it wouldn't behave for some reason, so, apologies for the "homemade" solution in Odoo shell.

This is tested on a production database that is having some performance issues, that's why we're doing a performance pass on a handful of modules

**Before:**

```
>>> from datetime import datetime as dt
>>> so = env["sale.order"].search([], limit=1)
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
3.842279 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.563739 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.670987 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.658426 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.532004 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.629248 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.555761 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.684823 seconds
```

**After**:

```
>>> from datetime import datetime as dt
>>> so = env["sale.order"].search([], limit=1)
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
3.460797 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.454272 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.428768 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.399137 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.413592 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.444747 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.380238 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.397538 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.412193 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.373276 seconds
>>> start = dt.now(); so._compute_product_assortment_ids(); print((dt.now() - start).total_seconds(), "seconds")
0.430472 seconds
```

To add to this, when pressing "create" from the interface, and the partner is still empty, the new function checks that the partner is not set and is not executed, whereas the old function still performs the full search.