@aleuffre 2024-12-10
Sorry, I made a new change, specifically

```py
                filters_partner_domain = filters_partner_domain.filtered(
                    lambda x, partner=partner: partner in x.partner_ids
                    or partner.filtered_domain(x._get_eval_partner_domain())
                )
```

This lets us skip computing `all_partner_ids` and the first invocation of the method goes down from 3.8 seconds to 1.8 seconds on my DB.
Also did some cleanup on the onchange below, though the impact in terms of performance is almost negligible.