@carlosdauden 2024-12-10
You need changes in product_assortment which I see are not at 14.0 either:
https://github.com/OCA/product-attribute/blob/15.0/product_assortment/models/ir_filters.py#L62-L71