@aleuffre 2024-12-10
> This version does not have the changes applied properly, this backport was made, but it was not quite right: #3135 as in addition to these changes there were others that did not carry over. Just take a look at the file in 15.0:
> 
> https://github.com/OCA/sale-workflow/blob/ad9d3f630683f6ba32ec29270b484f9c1255b2ad/sale_order_product_assortment/models/sale_order.py#L18-L38
> 
> Test the performance with these changes by simply copying and pasting.

I have vague memories of this PR. I think the problem, as you pointed out, was that there were changes not ported in another module in another repo, and due to the fact it can take a while to get PRs merged and we were in a hurry, it was decided to adapt the changes to the module as is.

Having to compute `all_partner_ids` just to see if a single partner is in a recordset of hundreds, if not thousands, every time, is really heavy, as you can imagine.

I'll check the PRs you mentioned.

Thank you for your time!