@aleuffre 2025-01-07
@carlosdauden Hello, sorry to bother you, but since you worked on the performance of this module, I figured it wouldn't hurt to ask for a suggestion.

Me and @francesco-ooops were running some more performance tests and while this PR helps a lot when the method in question is called (for example, when the SO is opened for the first time, or when the partner is changed), a lot of other operations (such as adding a SOL) are still very slow. And notably, no code in this module should really affect them. Removing this module makes SOs usable, whereas with this module (even with this PR) SOs are painfully slow, we're talking 10+ seconds to add a SOL and choose a product.

After a few tests, commenting parts of the view etc, I think I found the culprit. The field `allow_product_ids` in our case can contain over 35k records. When calling the onchange on the sale order line, the client sends the list of 35k ids to the server (that's ~~250kb of data!), and I assume the server reads it somehow because the onchange itself is painfully slow (over 5 seconds) even if the field `allow_product_ids` is not really used.

Commenting this part of the view

```xml
<attribute
                    name="domain"
                    operation="domain_add"
                    condition="parent.has_allowed_products"
                >[('id', 'in', parent.allowed_product_ids)]</attribute>
```
didn't help at all. Only when I also commented

```xml
                <field name="allowed_product_ids" invisible="1" />
```

I saw a noticeable performance gain.

Do you have any suggestions on how to proceed? Ideally I think I'd like to add the domain that is now used for the `.search()` in the compute directly to the view...