@aleuffre 2025-01-07
> Are you tested this in v15.0?

Not yet, but I will