@aleuffre 2025-01-08
> Are you tested this in v15.0?

I have tested it. From a clean DB with `sale_order_product_assortment`
1. Add 35k products (`for i in range(35000): env["product.product"].create({"name": i})`)
2. Create an assortment that includes all products, all partners (I used the domain `("id", ">", 0)`
3. Create a SO

Odoo is noticeably slow when selecting the partner, and when adding each Sales order line. Of course it's not quite as slow as a full production database with several other modules installed, but you can see the problem appearing there already.